Add compatibility with <iron-form>
This proposed change fixes PolymerElements/radio-button-group#66 and
PolymerElements/radio-button-group#58, by adding support for
`IronFormElementBehavior` and, additionally,
`IronFormValidatableBehavior`.

The end result is that calls to `<iron-form>.serialize()` return a
payload that includes a key for each `paper-radio-group` contained
within it (in addition to other fields), where the value at that key is
the value of `selected` (or, in other words, the value of
`attrForSelected` on the selected item).

Implementing `IronFormValidatableBehavior` means that `<iron-form>`
elements will also respect `allowEmptySelection` when validating forms
that contain `<paper-radio-group>`(s).

- <paper-radio-group> now includes the behaviors mentioned above
- in compliance with `IronFormElementBehavior`, `value` is provided as
  a synonym of the property `selected`
- in compliance with `IronFormValidatableBehavior`:
  - a new `validate` method exists to validate the value of a
    `<paper-radio-group>`
  - the property `required` now reflects the value of
    `allowEmptySelection`