Add compatibility with <iron-form>
This proposed change fixes PolymerElements/radio-button-group#58, by
adding support for `IronFormElementBehavior`.

The end result is that calls to `<iron-form>.serialize()` return a
payload that includes a key for each `paper-radio-group` contained
within it (in addition to other fields), where the value at that key is
the value of `selected` (or, in other words, the value of
`attrForSelected` on the selected item).

- <paper-radio-group> now includes the behavior mentioned above
- in compliance with that behavior, `value` is provided as a synonym of
  the property `selected`